requirements.txt
async-lru==1.0.2
boltons==20.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
emoji==1.2.0
Flask==1.1.2
greenlet==1.0.0
h11==0.12.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.2
pyaes==1.6.1
pydantic==1.8.1
Pyrogram==1.1.13
pyromod==1.5
PySocks==1.7.1
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
SQLAlchemy==1.4.2
starlette==0.13.6
TgCrypto==1.2.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
